+++ packages/components/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sl-design-system/shared
 
+## 0.1.5
+
+### Patch Changes
+
+- [#543](https://github.com/sl-design-system/components/pull/543) [`6f125e6`](https://github.com/sl-design-system/components/commit/6f125e6886f85fcfb8521e656f682e7385fe8aff) - Various improvements to data source & grid state
+
 ## 0.1.4
 
 ### Patch Changes