+++ packages/components/textarea/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sl-design-system/textarea
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [[`6f125e6`](https://github.com/sl-design-system/components/commit/6f125e6886f85fcfb8521e656f682e7385fe8aff)]:
+  - @sl-design-system/shared@0.1.5
+
 ## 0.1.4
 
 ### Patch Changes